backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pytest==7.4.3
pytest-cov==4.1.0
PyYAML==6.0.1
tiktoken==0.5.2
